Fix selected tab color token for Tabs (#886)
Addresses https://github.com/auth0/cosmos/issues/878

**Before:**
<img width="880" alt="screen shot 2018-09-03 at 11 31 05" src="https://user-images.githubusercontent.com/239215/44999086-21fabf80-af91-11e8-851e-91cd11ada8af.png">

**After:** 
<img width="896" alt="screen shot 2018-09-03 at 11 30 51" src="https://user-images.githubusercontent.com/239215/44999093-2aeb9100-af91-11e8-8f1e-3e0ee1966e52.png">

**Important:**
The token for the selected tab was not present anymore. So changed to use the correct one. This was actually caught in the **PoC environment** with the resets/globals off. So it's not reproducible in our Docs

Question: Should try to validate the tokens we use to avoid these errors @siddharthkp ?