Layouts: Row, Gallery, Columns and Page (#1282)
This PR is related to #1191 and introduces a bunch of important things:

## Margin Resets 

It abstract the layout margin reset so it removes margins from the children of the layouts. 

This is a temporary solution until we remove margins on v1 and @siddharthkp made a comment that the name of the prop sucks. He is right, I need help finding a new name.

## Gutters

It cleans up all layouts from CSS custom properties.

## Gallery options

It introduces gallery options to create different gallery sizes. I need help with the naming of this prop too :)

## Are we friends with the margin helper API?

Yes! We've tested the layouts on our Manage implementation using the margin prop and although we had to fight specificity  we conquer it by adding a class to the children of the layout (thanks @siddharthkp for the idea).

This Pr also adds an `!important` flag to the margin helper class.

## What reminds to be done:

- @landitus decide on the number of gutters and sizes.
- Add a `...prop` to all components
- Follow the same implementation for stack and app layouts