Automatically add id inside form fields (#1439)
Adds for + id pair for all form fields. (Fixes https://github.com/auth0/cosmos/issues/1438)

I also realised that clicking the label in the old API worked only for TextInput, TextArea and Select (not Switch, Radio, Checkbox, etc.)

You can test them here:

| API        | Stories |
| ------------- |-------------|
| Production new API | [all fields](https://auth0-cosmos.now.sh/sandbox/?selectedKind=Form&selectedStory=all%20the%20fields&full=0&addons=1&stories=1&panelRight=0) |
| Production Old API | [switch](https://auth0-cosmos.now.sh/sandbox/?selectedKind=Form&selectedStory=switch%20field%20-%20old%20API&full=0&addons=1&stories=1&panelRight=0) |
| This branch | [all fields](https://auth0-cosmos-sid-form-field-a11y.now.sh/sandbox/?selectedKind=Form&selectedStory=all%20the%20fields&full=0&addons=1&stories=1&panelRight=0) |


---

Please read the long story in the code comments and review comments: 
https://github.com/auth0/cosmos/pull/1439/files

This is definitely peak complexity. We need to decide if we want to own this complexity up or pass the responsibility to add `for + id` to the user instead.

---

I'm not sure what's the expectation when you click a label with a Radio and Checkbox groups 🤔